split blobs into two tests
