+++ test/blobs.js
@@ -6,6 +6,8 @@ var pull      = require('pull-stream')
 var u         = require('./util')
 var cont      = require('cont')
 var Hasher    = require('multiblob/util').createHash
+var createClient = require('../client')
+var ssbKeys   = require('ssb-keys')
 
 // create 3 servers
 // give them all pub servers (on localhost)
@@ -19,8 +21,45 @@ function read (filename) {
   return toPull.source(fs.createReadStream(filename))
 }
 
-tape('replicate blobs between 2 peers - explicit want request', function (t) {
+tape('a client can request a blob', function (t) {
 
+  var u = require('./util')
+
+  var sbotA = u.createDB('test-blobs-alice0', {
+      port: 45450, host: 'localhost', timeout: 1000,
+    }).use(blobs)
+
+  var alice = sbotA.feed
+
+  pull(
+    read(path.join(__filename)),
+    sbotA.blobs.add(function (err, hash) {
+      if(err) throw err
+      createClient(sbotA.feed.keys, sbotA.manifest)({
+        port: 45450, key: sbotA.feed.keys.public
+      }, function (err, rpc) {
+
+        rpc.blobs.has(hash, function (err) {
+          if(err) throw err
+          pull(
+            rpc.blobs.get(hash),
+            pull.collect(function (err, ary) {
+              if(err) throw err
+              var data = Buffer.concat(ary.map(function (e) {
+                return new Buffer(e, 'base64')
+              }))
+              sbotA.close()
+              t.equal(ssbKeys.hash(data), hash)
+              t.end()
+            })
+          )
+        })
+      })
+    })
+  )
+})
+
+tape('replicate blobs between 2 peers - explicit want request', function (t) {
   var u = require('./util')
 
   var sbotA = u.createDB('test-blobs-alice1', {
@@ -38,15 +77,18 @@ tape('replicate blobs between 2 peers - explicit want request', function (t) {
 
   pull(
     read(path.join(__filename)),
-    sbotA.blobs.add()
+    sbotA.blobs.add(function (err, hash) {
+      if(err) throw err
+    })
   )
 
   sbotA.on('blobs:got', function (hash) {
+    console.log('BLOBS', hash)
     console.log('added', hash)
     sbotB.blobs.want(hash, function (err) {
-      console.log('has!', err, hash)
       if(err) throw err
-      sbotB.blobs.has(hash, function (_, has) {
+      sbotB.blobs.has(hash, function (err, has) {
+        if(err) throw err
         t.ok(has)
         t.end()
         sbotA.close()
@@ -60,7 +102,6 @@ tape('replicate blobs between 2 peers - explicit want request', function (t) {
 })
 
 tape('replicate published blobs between 2 peers', function (t) {
-
   var sbotA = u.createDB('test-blobs-alice2', {
       port: 45451, host: 'localhost', timeout: 1000,
     }).use(gossip).use(friends).use(replicate).use(blobs)
@@ -92,6 +133,7 @@ tape('replicate published blobs between 2 peers', function (t) {
       // and then emit this event.
 
       sbotB.on('blobs:got', function (_hash) {
+        console.log("BLOBS GOT", _hash)
         t.equal(_hash, hash)
         sbotB.blobs.has(hash, function (err, okay) {
           t.ok(okay, 'file replicated:' + hash)
@@ -104,192 +146,4 @@ tape('replicate published blobs between 2 peers', function (t) {
   )
 })
 
-tape('avoid flooding a peer with blob requests', function (t) {
 
-  var sbotA = u.createDB('test-blobs-alice3', {
-      port: 45451, host: 'localhost', timeout: 1000,
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var alice = sbotA.feed
-
-  var sbotB = u.createDB('test-blobs-bob3', {
-      port: 45452, host: 'localhost', timeout: 1000,
-      seeds: [{port: 45451, host: 'localhost', key: sbotA.feed.keys.public}]
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var bob = sbotB.feed
-
-  var hasher = Hasher()
-
-  sbotA.on('blobs:has', function (r) {
-    console.log('REQUEST', r)
-  })
-
-  pull(
-    read(__filename),
-    hasher,
-    pull.drain(null, function (err) {
-
-      var hash = hasher.digest
-      console.log('WANT:', hash)
-
-      cont.para([
-        alice.add({type: 'post', text: 'this file', js: {ext: hash}}),
-        alice.add({type: 'contact', following: true, contact: { feed: bob.id }}),
-        bob.add({type: 'contact', following: true, contact: {feed: alice.id}})
-      ])(function (err, data) {
-        if(err) throw err
-      })
-      // bob should not request `hash` more than once.
-
-      t.plan(1)
-
-      sbotA.on('blobs:has', function (h) {
-        console.log('HAS', h)
-        t.deepEqual(h, [hash])
-      })
-
-      sbotB.once('rpc:authorized', function (rpc) {
-        console.log('rpc:authorized')
-        rpc.on('closed', function () {
-          console.log('CLOSE')
-          t.end()
-          sbotA.close()
-          sbotB.close()
-        })
-      })
-    })
-  )
-})
-
-
-tape('request missing blobs again after reconnect', function (t) {
-
-  var sbotA = u.createDB('test-blobs-alice4', {
-      port: 45453, host: 'localhost', timeout: 1000,
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var alice = sbotA.feed
-
-  var sbotB = u.createDB('test-blobs-bob4', {
-      port: 45454, host: 'localhost', timeout: 1000,
-      seeds: [{port: 45453, host: 'localhost', key: sbotA.feed.keys.public}]
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var bob = sbotB.feed
-
-  var hasher = Hasher()
-
-  sbotA.on('blobs:has', function (r) {
-    console.log('REQUEST', r)
-  })
-
-  pull(
-    read(__filename),
-    hasher,
-    pull.drain(null, function (err) {
-
-      var hash = hasher.digest
-      console.log('WANT:', hash)
-
-      cont.para([
-        alice.add({type: 'post', text: 'this file', js: {ext: hash}}),
-        alice.add({type: 'contact', following: true, contact: { feed: bob.id }}),
-        bob.add({type: 'contact', following: true, contact: {feed: alice.id}})
-      ])(function (err, data) {
-        if(err) throw err
-      })
-      // bob should not request `hash` more than once.
-
-      t.plan(2)
-
-      sbotA.on('blobs:has', function (h) {
-        console.log('HAS', h)
-        t.deepEqual(h, [hash])
-      })
-
-      sbotB.once('rpc:authorized', function (rpc) {
-        console.log('rpc:authorized - 1')
-        sbotB.once('rpc:authorized', function (rpc) {
-          console.log('rpc:authorized - 2')
-          rpc.on('closed', function () {
-            console.log('CLOSE')
-            sbotA.close()
-            sbotB.close()
-            t.end()
-          })
-        })
-      })
-    })
-  )
-})
-
-tape('emit "has" event to let peer know you have blob now', function (t) {
-
-  var sbotA = u.createDB('test-blobs-alice5', {
-      port: 45455, host: 'localhost', timeout: 1000,
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var alice = sbotA.feed
-
-  var sbotB = u.createDB('test-blobs-bob5', {
-      port: 45456, host: 'localhost', timeout: 1000,
-      seeds: [{port: 45455, host: 'localhost', key: sbotA.feed.keys.public}]
-    }).use(gossip).use(friends).use(replicate).use(blobs)
-
-  var bob = sbotB.feed
-
-  var hasher = Hasher()
-
-  sbotA.on('blobs:has', function (r) {
-    console.log('REQUEST', r)
-  })
-
-  pull(
-    read(__filename),
-    hasher,
-    pull.drain(null, function (err) {
-
-      var hash = hasher.digest
-      console.log('WANT:', hash)
-
-      cont.para([
-        alice.add({type: 'post', text: 'this file', js: {ext: hash}}),
-        alice.add({type: 'contact', following: true, contact: { feed: bob.id }}),
-        bob.add({type: 'contact', following: true, contact: {feed: alice.id}})
-      ])(function (err, data) {
-        if(err) throw err
-      })
-      // bob should not request `hash` more than once.
-
-      t.plan(2)
-
-      sbotB.on('blobs:got', function (h) {
-        t.equal(h, hash)
-        sbotA.close()
-        sbotB.close()
-        t.end()
-      })
-
-      //wait for bob to request the hash
-      //then add that file.
-      sbotA.on('blobs:has', function (h) {
-        console.log('HAS', h)
-        t.deepEqual(h, [hash])
-
-        pull(
-          read(__filename),
-          sbotA.blobs.add(null, function (err, hash) {
-            //have now added the blob to 
-          })
-        )
-      })
-    })
-  )
-
-  //this test should only require one connection.
-  var n = 0
-  sbotB.on('rpc:authorized', function (rpc) {
-    if(++n > 1) throw new Error('connected twice')
-  })
-})
